+++ qthriftd/qthrift_bgp_updater.c
@@ -37,7 +37,7 @@
  */
 gboolean
 qthrift_bgp_updater_on_update_push_route (const protocol_type p_type, const gchar * rd, const gchar * prefix, const gint32 prefixlen, 
-                                          const gchar * nexthop, const gint32 ethtag, const gchar * esi, const gchar * macaddress, 
+                                          const gchar * nexthop, const gint64 ethtag, const gchar * esi, const gchar * macaddress,
                                           const gint32 l3label, const gint32 l2label, const gchar * routermac)
 {
   GError *error = NULL;
@@ -53,7 +53,7 @@ qthrift_bgp_updater_on_update_push_route (const protocol_type p_type, const gcha
   if(IS_QTHRIFT_DEBUG_NOTIFICATION)
     {
     char ethtag_str[20];
-    sprintf(ethtag_str,"ethtag %u", ethtag);
+    sprintf(ethtag_str,"ethtag %ld", ethtag);
 
     zlog_info ("onUpdatePushRoute(rd %s, pfx %s, nh %s, l3label %d, l2label %d, %s%s, %s%s, %s %s%s) sent %s",
              rd, prefix? prefix:"none", nexthop, l3label, l2label,
@@ -72,7 +72,7 @@ qthrift_bgp_updater_on_update_push_route (const protocol_type p_type, const gcha
  */
 gboolean
 qthrift_bgp_updater_on_update_withdraw_route (const protocol_type p_type, const gchar * rd, const gchar * prefix, const gint32 prefixlen, const gchar * nexthop,  
-                                              const gint32 ethtag, const gchar * esi, const gchar * macaddress, const gint32 l3label, const gint32 l2label)
+                                              const gint64 ethtag, const gchar * esi, const gchar * macaddress, const gint32 l3label, const gint32 l2label)
 {
   GError *error = NULL;
   gboolean response;
@@ -87,7 +87,7 @@ qthrift_bgp_updater_on_update_withdraw_route (const protocol_type p_type, const 
   if(IS_QTHRIFT_DEBUG_NOTIFICATION)
     {
       char ethtag_str[20];
-      sprintf(ethtag_str,"ethtag %u", ethtag);
+      sprintf(ethtag_str,"ethtag %ld", ethtag);
 
       zlog_info ("onUpdateWithdrawRoute(rd %s, pfx %s/%d, nh %s, label %d, l2label %d, %s%s %s%s %s) sent %s", \
                 rd, prefix? prefix:"none", prefixlen, nexthop, l3label, l2label,
