+++ qthriftd/vpnservice.thrift
@@ -139,7 +139,7 @@ enum protocol_type {
       *             ignored if L3VPN-MPLS is configured.
       */
      i32 pushRoute(1:protocol_type p_type, 2:string prefix, 3:string nexthop, 4:string rd,
-                   5:i32 ethtag, 6:string esi, 7:string macaddress,
+                   5:i64 ethtag, 6:string esi, 7:string macaddress,
                    8:i32 l3label, 9:i32 l2label, 10:encap_type enc_type, 11:string routermac),
      /*
       * 'p_type' is mandatory
@@ -147,7 +147,7 @@ enum protocol_type {
       * label (v4LU) as a string (eg: "2500")
       */
      i32 withdrawRoute(1:protocol_type p_type, 2:string prefix, 3:string rd,
-                       4:i32 ethtag, 5:string esi, 6:string macaddress),
+                       4:i64 ethtag, 5:string esi, 6:string macaddress),
      i32 setEbgpMultihop(1:string peerIp, 2:i32 nHops),
      i32 unsetEbgpMultihop(1:string peerIp),
      i32 setUpdateSource(1:string peerIp, 2:string srcIp),
@@ -181,11 +181,11 @@ enum protocol_type {
    // 'p_type' is mandatory. indicates the origin of data
    oneway void onUpdatePushRoute(1:protocol_type p_type, 2:string rd, 3:string prefix, 
                                  4:i32 prefixlen, 5:string nexthop, 
-                                 6:i32 ethtag, 7:string esi, 8:string macaddress,
+                                 6:i64 ethtag, 7:string esi, 8:string macaddress,
                                  9:i32 l3label, 10:i32 l2label, 11:string routermac),
    oneway void onUpdateWithdrawRoute(1:protocol_type p_type, 2:string rd, 3:string prefix, 
                                      4:i32 prefixlen, 5:string nexthop,
-                                     6:i32 ethtag, 7:string esi, 8:string macaddress,
+                                     6:i64 ethtag, 7:string esi, 8:string macaddress,
                                      9:i32 l3label, 10:i32 l2label),
    // tell them we're open for business
    oneway void onStartConfigResyncNotification(),
